requirements.txt
aiosmtpd==1.4.6
amqp==5.3.1
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
Brotli==1.1.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.7
cssselect2==0.8.0
Django==5.2.5
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
flower==2.0.1
fonttools==4.59.2
html5lib==1.1
humanize==4.13.0
idna==3.10
inflection==0.5.1
kombu==5.5.4
lxml==6.0.1
oscrypto==1.3.0
packaging==25.0
pillow==11.3.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
psycopg2-binary
pycparser==2.22
pydyf==0.11.0
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
PyJWT==2.10.1
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
reportlab==4.4.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.5.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
